fixed* health resetting to 20 on login
*technically it's just a bandaid fix, because of how trinkets work, so this is the simplest "fix"

A known unintended effects of this is that if you have more than 20hp, but less than your maximum, removing your trinkets and reequipping it will heal you to your maximum hp. The same applies with logging off and on. I'm not gonna even try to fix it, so deal with it